+++ test/test-programs/defer-abort.js
@@ -1,0 +1,1 @@
+setImmediate(function doExplicitAbort() { process.abort(); });